backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-security-too
flask-restful
werkzeug
bcrypt
argon2-cffi
